pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "magnetopy"
version = "0.1.1"
description = "Magnetism in Python"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "matplotlib",
    "scikit-learn",
    "GitPython",
]

[project.optional-dependencies]
dev = [
    "black",
    "pytest",
    "pytest-cov",
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocstrings-python",
    "markdown-katex",
    "mkdocs-jupyter",
]

[project.scripts]
mp-calibration = "magnetopy.cli.calibration_install:main"

[tool.setuptools]
packages = ["magnetopy"]

[project.urls]
"Homepage" = "https://github.com/RinehartGroup/MagnetoPy.git"